docker-example/python-client/requirements.txt
# Python Client Dependencies
# Install these packages before running the pipeline

requests>=2.31.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
